Conflicts:
	sql_mr/mr00023_mangos_spell_bonus_data.sql
	src/game/AchievementMgr.cpp
	src/game/BattleGround.cpp
	src/game/BattleGround.h
	src/game/BattleGroundDS.cpp
	src/game/BattleGroundMgr.cpp
	src/game/BattleGroundRV.cpp
	src/game/BattleGroundSA.cpp
	src/game/BattleGroundSA.h
	src/game/Creature.cpp
	src/game/CreatureAI.h
	src/game/DBCfmt.h
	src/game/GameObject.cpp
	src/game/GridNotifiers.h
	src/game/Map.cpp
	src/game/Map.h
	src/game/Object.cpp
	src/game/Player.cpp
	src/game/QueryHandler.cpp
	src/game/Spell.cpp
	src/game/SpellAuras.cpp
	src/game/SpellEffects.cpp
	src/game/SpellHandler.cpp
	src/game/SpellMgr.cpp
	src/game/Unit.cpp
	src/game/UnitAuraProcHandler.cpp
	src/game/Vehicle.cpp
	src/game/debugcmds.cpp
	win/VC90/game.vcproj

+++ src/game/WaypointMovementGenerator.cpp
@@ -326,9 +326,9 @@ void FlightPathMovementGenerator::Reset(Player & player)
 
 bool FlightPathMovementGenerator::Update(Player &player, const uint32 &diff)
 {
-    int32 pointId = player.movespline->currentPathIdx();
+    uint32 pointId = (uint32)player.movespline->currentPathIdx();
     // currentPathIdx returns lastIdx + 1 at arrive
-    while (i_currentNode < pointId)
+    while (i_currentNode < (int)pointId)
     {
         DoEventIfAny(player,(*i_path)[i_currentNode],true);
         DoEventIfAny(player,(*i_path)[i_currentNode],false);
